// src/app/models/user-registration-dto.model.ts
export interface UserRegistrationDto {
    fullName: string;
    bloodGroup: string;
    userId: string;
    mobileNumber: string;
    emailId: string;
    password: string;
    country: string;
    state: string;
    district: string;
    city: string;
    availabilityToDonateBlood: boolean;
  }
  